rename route for getting a specific post by guid from post/:guid to posts/:guid, update all corresponding tests and frontend
