+++ src/backend/web/routes/index.js
@@ -4,7 +4,6 @@ const path = require('path');
 const admin = require('./admin');
 const opml = require('./opml');
 const planet = require('./planet');
-const post = require('./post');
 const posts = require('./posts');
 const sse = require('./server-sent-events');
 const login = require('./login');
@@ -16,7 +15,6 @@ router.use(express.static(path.join(__dirname, '../../../frontend')));
 router.use('/admin', admin);
 router.use('/opml', opml);
 router.use('/planet', planet);
-router.use('/post', post);
 router.use('/posts', posts);
 router.use('/feed-updates', sse);
 router.use('/login', login);
